+++ src/main/java/com/example/demo/service/CustomerService.java
@@ -25,14 +25,14 @@ public class CustomerService {
 
 	public List<Customer> getAllCustomers() { return customerDao.selectAllCustomers(); }
 
-	public Optional<Customer> getCustomerById(Long id) { return customerDao.selectCustomerById(id); }
+	public Optional<Customer> getCustomerById(int id) { return customerDao.selectCustomerById(id); }
 
 	public Optional<Customer> getCustomerByUsername(String username) {
 		return customerDao.selectCustomerByUsername(username);
 	}
 
-	public void deleteCustomer(Long id) { customerDao.deleteCustomerById(id); }
+	public void deleteCustomer(int id) { customerDao.deleteCustomerById(id); }
 
-	public Customer updateCustomer(Long id, Customer newCustomer) { return customerDao.updateCustomerById(id, newCustomer); }
+	public Customer updateCustomer(int id, Customer newCustomer) { return customerDao.updateCustomerById(id, newCustomer); }
 
 }
